requirements.txt
termcolor==2.4.0
openai==1.10.0
PyYAML==6.0.1
streamlit==1.31.0
st-gsheets-connection==0.0.4
